refactor: centralize GitHub conclusion contracts
